% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_crit_solver.R
\name{f_crit_solver}
\alias{f_crit_solver}
\title{Critical Intermittency Calculator}
\usage{
f_crit_solver(
  D,
  Omega,
  cement_thickness_ratio = 0.04,
  k = 10^1.11,
  n = 2.26,
  rho_s = 2800,
  M_min = 100.0869,
  rho_f = 1025,
  H = 1,
  nu = 9.37 * 10^-7,
  L_dune = 28,
  H_dune = 0.4,
  H_swave = 0.2,
  T_pwave = 10,
  L_wave = 10
)
}
\arguments{
\item{D}{ooid diameter in m}

\item{Omega}{\eqn{\Omega}, the CaCO\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}} mineral saturation state, typically with respect to aragonite or calcite}

\item{cement_thickness_ratio}{parameter that scales the thickness of cement needed to form the hardground; this factor is multiplied by the grain diameter, so a value of 1 means that a thickness = grain diameter is required, while a value of 0.01 means that a thickness of 1\% of the grain diameter is required}

\item{k}{CaCO\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}} precipitation rate constant in umol/m\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}/hr, this is sensitive to temperature and mineralogy and can be set with an optional helper function}

\item{n}{CaCO\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}} precipitation reaction order (unitless), this is also sensitive to temperature and mineralogy and can be set with an optional helper function}

\item{rho_s}{density of sediment in kg/m\if{html}{\out{<sup>}}3\if{html}{\out{</sup>}}, this is sensitive to mineralogy}

\item{M_min}{molar mass of mineral (g/mol), default value is for CaCO\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}}}

\item{rho_f}{density of fluid (kg/m\if{html}{\out{<sup>}}3\if{html}{\out{</sup>}}), this is sensitive to temperature and salinity, helper function for calculating seawater density is available in Oomega Toolbox package}

\item{H}{water depth in m}

\item{nu}{kinematic viscosity of fluid in m\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}/s, this is sensitive to temperature and salinity, helper function for calculating dynamic viscosity for seawater is available in Oomega Toolbox package; kinematic viscosity can be calculated from dynamic viscosity and density}

\item{L_dune}{characteristic dune wavelength (m)}

\item{H_dune}{characteristic dune height (m)}

\item{H_swave}{significant wave height (m)}

\item{T_pwave}{peak wave period (s)}

\item{L_wave}{wave length (m)}
}
\value{
critical intermittency (unitless)
}
\description{
Calculates the critical intermittency, below which ooids will cement together rather than continue to grow
}
\details{
This function uses a bedload transport equation and an estimate of the cement thickness necessary to cement grains together during transport to estimate the minimum intermittency, below which ooids will be cemented together and can no longer continue to grow as individual coated grains. Field observations suggest that the critical cement thickness ratio is 2-4\% of the grain diameter (cement_thickness_ratio = 0.02 to 0.04) for sand-sized marine ooids in Ambergris shoal.
}
\examples{
f_crit1 <- f_crit_solver(500*10^-6,5)
f_crit2 <- mapply(f_crit_solver,D=500*10^-6,Omega = seq(from=2,to=10,by=0.5))
f_crit3 <- mapply(f_crit_solver,D=seq(from=400,to=500,by=10)*10^-6,Omega=5)

}
