% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OomegaSolver.R
\name{OomegaSolver}
\alias{OomegaSolver}
\title{Oomega Solver}
\usage{
OomegaSolver(
  D,
  k = 10^1.11,
  n = 2.26,
  rho_s = 2800,
  M_min = 100.0869,
  rho_f = 1025,
  nu = 9.37 * 10^-7,
  H = 1,
  intermittency = 0.15
)
}
\arguments{
\item{D}{ooid diameter in m}

\item{k}{CaCO\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}} precipitation rate constant in umol/m\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}/hr, this is sensitive to temperature and mineralogy and can be set with an optional helper function}

\item{n}{CaCO\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}} precipitation reaction order (unitless), this is also sensitive to temperature and mineralogy and can be set with an optional helper function}

\item{rho_s}{density of sediment in kg/m\if{html}{\out{<sup>}}3\if{html}{\out{</sup>}}, this is sensitive to mineralogy}

\item{M_min}{molar mass of mineral (g/mol), default value is for CaCO\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}}}

\item{rho_f}{density of fluid (kg/m\if{html}{\out{<sup>}}3\if{html}{\out{</sup>}}), this is sensitive to temperature and salinity, helper function for calculating seawater density is available in Oomega Toolbox package}

\item{nu}{kinematic viscosity of fluid in m\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}/s, this is sensitive to temperature and salinity, helper function for calculating dynamic viscosity for seawater is available in Oomega Toolbox package; kinematic viscosity can be calculated from dynamic viscosity and density}

\item{H}{water depth in m}

\item{intermittency}{intermittency of movement, must be in the range 0-1 where 0 implies no movement and 1 implies constant movement}
}
\value{
calculated \eqn{\Omega} value
}
\description{
Calculates the CaCO\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}} saturation state (\eqn{\Omega}) value at which precipitation and abrasion rates are balanced for the given ooid diameter.
}
\details{
This function is an R implementation of the OomegaSolver function described in the following references:

Ingalls, M., Fetrow, A. C., Snell, K. E., Frantz, C. M., & Trower, E. J. (2022). Lake level controls the recurrence of giant stromatolite facies. Sedimentology, 69(4), 1649–1674.

Trower, E. J., Smith, B. P., Koeshidayatullah, A. I., & Payne, J. L. (2022). Marine ooid sizes record Phanerozoic seawater carbonate chemistry. Geophysical Research Letters, 49(22), e2022GL100800.

The basic premise is to use measurements of ooid diameter (which may need to be corrected if measured from thin sections) to reconstruct the carbonate chemistry of the waters they formed in.
}
\examples{
Omega <- OomegaSolver(D = 500*10^-6,
                      H = 2,
                      intermittency = 0.1)
                      
Omega <- mapply(OomegaSolver, 
                D = c(400,500,600)*10^-6, 
                H = 1.5)

}
