% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seawaterdensity.R
\name{seawaterdensity}
\alias{seawaterdensity}
\title{Seawater Density Calculator}
\usage{
seawaterdensity(tempC, sal)
}
\arguments{
\item{tempC}{temperature in degrees Celsius; valid for -2 to 40 degrees C}

\item{sal}{practical salinity in g/kg; valid for 0 to 42 g/kg}
}
\value{
density in kg/m^3
}
\description{
Calculates seawater density as a function of temperature in Celsius and salinity in g/kg
}
\details{
This is a function from F. J. Millero, and A. Poisson,
International one-atmosphere equation of state of seawater.
Deep-Sea Research, 28A (6), 625 – 629, 1981.
}
\examples{
rho_sw1 <- seawaterdensity(25,35)
rho_sw2 <- seawaterdensity(c(25,30,35),35)

}
